refactored to adapt the code of the delete method to the database -> I created a service to handle deletions

+++ src/main/java/com/songifyDatabase/song/infructure/controller/SongMapper.java
@@ -28,7 +28,7 @@ public class SongMapper {
     public static CreateSongResponseDto mapFromSongToCreateSongResponseDto(Song song) {
         return new CreateSongResponseDto(song);
     }
-    public static DeleteSongResponseDto mapFromSongToDeleteSongResponseDto(Integer id) {
+    public static DeleteSongResponseDto mapFromSongToDeleteSongResponseDto(Long id) {
         return new DeleteSongResponseDto("you deleted songName with id: " + id, HttpStatus.OK);
     }
